+++ nnetsauce/datasets/__init__.py
@@ -1,0 +1,3 @@
+from .dowload import Downloader
+
+__all__ = ["Downloader"]